+++ src/button/HISTORY.md
@@ -4,7 +4,8 @@ Button Change History
 @VERSION@
 ------
 
-* No changes.
+* Added a `labelHTML` attribute to `Y.ButtonCore` for nested HTML label support
+* Due to a fix in `Y.Widget` (#1125), `Y.Button` now correctly retains all node attributes upon render
 
 3.12.0
 ------